improving command documentation
